Update unwind_stubs.c after 36dbabc1a9728cf7843428360458ae5842fbb7fa

+++ src/unwind_stubs.c
@@ -7,13 +7,13 @@ _Unwind_Reason_Code _Unwind_ForcedUnwind (struct _Unwind_Exception * e,
 						 _Unwind_Stop_Fn f, void *p) { return _URC_NO_REASON; }
 void _Unwind_Resume (struct _Unwind_Exception * e) { }
 void _Unwind_DeleteException (struct _Unwind_Exception *e) { }
-unsigned long _Unwind_GetGR (struct _Unwind_Context *e, int i) { return 0; }
-void _Unwind_SetGR (struct _Unwind_Context * c, int i, unsigned long l) { }
-unsigned long _Unwind_GetIP (struct _Unwind_Context *g) { return 0; }
-unsigned long _Unwind_GetIPInfo (struct _Unwind_Context *c, int *i) { return 0; }
-void _Unwind_SetIP (struct _Unwind_Context *c, unsigned long l) { }
-unsigned long _Unwind_GetLanguageSpecificData (struct _Unwind_Context *c) { return 0; }
-unsigned long _Unwind_GetRegionStart (struct _Unwind_Context *c) { return 0; }
+uintptr_t _Unwind_GetGR (struct _Unwind_Context *e, int i) { return 0; }
+void _Unwind_SetGR (struct _Unwind_Context * c, int i, uintptr_t l) { }
+uintptr_t _Unwind_GetIP (struct _Unwind_Context *g) { return 0; }
+uintptr_t _Unwind_GetIPInfo (struct _Unwind_Context *c, int *i) { return 0; }
+void _Unwind_SetIP (struct _Unwind_Context *c, uintptr_t l) { }
+uintptr_t _Unwind_GetLanguageSpecificData (struct _Unwind_Context *c) { return 0; }
+uintptr_t _Unwind_GetRegionStart (struct _Unwind_Context *c) { return 0; }
 
 _Unwind_Reason_Code
 	  _Unwind_Resume_or_Rethrow (struct _Unwind_Exception *e) { return _URC_NO_REASON; }
